fix: type collision between DIDComm message class and Message namespace which will mess up our transition to workspaces Signed-off-by: Francisco Javier Ribo Labrador <elribonazo@gmail.com>
